feat: Add Sales overview to the application
Added a new route /overview to the App.js file to allow navigation to a new Sales overview page. -Also, modified the state/api.js file to include a new API endpoint to fetch sales data. Additionally, a new controller was created in the server/controllers/sales.js file to handle the endpoint request. Finally, the server/models/ProductStat.js file was modified to include an array of daily data instead of a single object.